+++ tests/testthat/test-tidyverse.R
@@ -23,13 +23,11 @@ test_that("use_tidy_dependencies() isn't overly informative", {
   expect_snapshot(use_tidy_dependencies())
 })
 
-test_that("use_tidy_eval() inserts the template file and Imports rlang", {
+test_that("use_tidy_eval() is defunct", {
   skip_if_not_installed("roxygen2")
 
   pkg <- create_local_package()
-  use_tidy_eval()
-  expect_match(dir_ls(proj_path("R")), "utils-tidy-eval.R")
-  expect_match(desc::desc_get("Imports", pkg), "rlang")
+  expect_snapshot(use_tidy_eval(), error = TRUE)
 })
 
 test_that("use_tidy_GITHUB-STUFF() adds and Rbuildignores files", {